Move Pandas related code outside of the model class
